+++ R/nitrogen_efficiency.R
@@ -16,8 +16,8 @@
 #' @param B_FERT_NORM_FR (numeric) The fraction of the application norm utilized
 #' 
 #' @examples 
-#' calc_n_efficiency(1019,'dekzand','GtIV','Zuidwest-Brabant',4.5,3.5,0.8,0.6,0.2,78,FALSE,1)
-#' calc_n_efficiency(256,'veen','GtII','Centraal Veehouderijgebied',4.5,3.5,0.8,0.6,0.2,250,FALSE,1)
+#' calc_n_efficiency(1019,'dekzand','IV','Zuidwest-Brabant',4.5,3.5,0.8,0.6,0.2,78,FALSE,1)
+#' calc_n_efficiency(256,'veen','II','Centraal Veehouderijgebied',4.5,3.5,0.8,0.6,0.2,250,FALSE,1)
 #'
 #' @return 
 #' The estimated index for the nitrogen use efficiency, as being affected by soil properties. A numeric value.
@@ -27,7 +27,7 @@ calc_n_efficiency <- function(B_LU_BRP, B_SOILTYPE_AGR, B_GWL_CLASS, B_AER_CBS, 
                               D_PBI, D_K, D_PH_DELTA, leaching_to, M_GREEN = FALSE, B_FERT_NORM_FR = 1){
   
   # add visual bindings
-  id = crops.obic = soils.obic = leaching_to_set = crop_catergory = bodem = gewas = B_GT = NULL
+  id = crops.obic = soils.obic = leaching_to_set = crop_catergory = bodem = gewas = NULL
   nf = n_eff = nf_sand.other = nf_sand.south = nf_clay = nf_peat = nf_loess = NUE = NULL
   soiltype = soiltype.n = croptype.nleach = crop_category = deposition = crop_waterstress = NULL
   I_P = I_K = I_PH = cf_pkph = decomposition = n_app = catchcrop = n_sp = D_NLEACH = NULL
@@ -42,8 +42,11 @@ calc_n_efficiency <- function(B_LU_BRP, B_SOILTYPE_AGR, B_GWL_CLASS, B_AER_CBS, 
   checkmate::assert_character(B_SOILTYPE_AGR, any.missing = FALSE, len = arg.length)
   checkmate::assert_subset(B_SOILTYPE_AGR, choices = unique(OBIC::soils.obic$soiltype))
   checkmate::assert_character(B_GWL_CLASS,any.missing = FALSE, len = arg.length)
-  checkmate::assert_subset(B_GWL_CLASS, choices = c('GtI','GtII','GtIII','GtIV','GtV','GtVI','GtVII','GtVIII',
-                                                  'GtIIb','GtIIIb','GtVb'), empty.ok = FALSE)
+  checkmate::assert_subset(B_GWL_CLASS, choices =c(
+    "I", "Ia", "Ic", "II", "IIa", "IIb", "IIc", "III", "IIIa", "IIIb", "IV",
+    "IVc", "IVu", "sV", "sVb", "V", "Va", "Vad", "Vao", "Vb", "Vbd", "Vbo", "VI", 
+    "VId", "VII", "VIId", "VIII", "VIIId", "VIIIo", "VIIo", "VIo"
+  ), empty.ok = FALSE)
   checkmate::assert_character(B_AER_CBS, any.missing = FALSE, min.len = 1, len = arg.length)
   checkmate::assert_subset(B_AER_CBS, choices = c('Zuid-Limburg','Zuidelijk Veehouderijgebied','Zuidwest-Brabant',
                                                   'Zuidwestelijk Akkerbouwgebied','Rivierengebied','Hollands/Utrechts Weidegebied',
@@ -90,9 +93,9 @@ calc_n_efficiency <- function(B_LU_BRP, B_SOILTYPE_AGR, B_GWL_CLASS, B_AER_CBS, 
   dt[crop_category == "grasland" , croptype.nleach := "gras"]
   
   # merge fraction of N leaching into 'dt', based on soil type x crop type x grondwatertrap
-  dt <- merge(dt, nleach_table[, list(bodem, gewas, B_GT, nf)], 
+  dt <- merge(dt, nleach_table[, list(bodem, gewas, B_GWL_CLASS, nf)], 
               by.x = c("soiltype.n", "croptype.nleach", "B_GWL_CLASS"), 
-              by.y = c("bodem", "gewas", "B_GT"), sort = FALSE, all.x = TRUE)
+              by.y = c("bodem", "gewas", "B_GWL_CLASS"), sort = FALSE, all.x = TRUE)
   
   # select the allowed effective N dose (in Dutch: N-gebruiksnorm), being dependent on soil type and region
   sand.south <- c('Zuid-Limburg','Zuidelijk Veehouderijgebied','Zuidwest-Brabant')